+++ led-display/led-display-web/led-display-web-spring-display/src/main/java/me/hmasrafchi/leddisplay/consumer/infrastructure/jpa/MatrixRepositorySpringJpa.java
@@ -3,6 +3,8 @@
  */
 package me.hmasrafchi.leddisplay.consumer.infrastructure.jpa;
 
+import java.math.BigInteger;
+
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
@@ -13,6 +15,6 @@ import me.hmasrafchi.leddisplay.consumer.data.jpa.MatrixJpa;
  *
  */
 @Repository
-interface MatrixRepositorySpringJpa extends JpaRepository<MatrixJpa, Integer> {
+interface MatrixRepositorySpringJpa extends JpaRepository<MatrixJpa, BigInteger> {
 
 }
